Significant fixes to resumption of discovery on Android.
Mostly fixes on ensuring services shutdown and startup cleanly. There's also a few fixes around finding proper network interfaces to join the multicast group on. This fixes an issue I was seeing on Self AP mode. Logging is verbose right now but I find it nice to see it's working.